pyproject.toml
[build-system]
requires = [ "poetry-core",]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "python-jobspy"
version = "1.1.82"
description = "Job scraper for LinkedIn, Indeed, Glassdoor, ZipRecruiter & Bayt"
authors = ["Cullen Watson <cullen@cullenwatson.com>", "Zachary Hampton <zachary@zacharysproducts.com>"]
homepage = "https://github.com/cullenwatson/JobSpy"
readme = "README.md"
keywords = [ "jobs-scraper", "linkedin", "indeed", "glassdoor", "ziprecruiter", "bayt", "naukri"]
[[tool.poetry.packages]]
include = "jobspy"

[tool.black]
line-length = 88

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
pandas = "^2.1.0"
numpy = ">=1.26.0"
pydantic = "^2.3.0"
tls-client = "^1.0.1"
markdownify = "^1.1.0"
regex = "^2024.4.28"
python-docx = "^1.1.0"
google-generativeai = "^0.8.0"
google-genai = "^1.0.0"
psycopg2-binary = "^2.9.0"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
black = "*"
pre-commit = "*"
